changes float to float32 or float64 as apparently float doesnt default to anything on mingw-go
removed dependencies on draw gocairo packages to symplify compilation on
windows

+++ synstation/init.go
@@ -44,7 +44,7 @@ func init() {
 
 			for k := 1; k <= overlapN; k++ {
 
-				fac := 1.0 - float(k)*(1.0-roverlap)
+				fac := 1.0 - float64(k)*(1.0-roverlap)
 
 				SystemChan[i].coIntC[overlapN-k].c = i - k
 				SystemChan[i].coIntC[overlapN-k].factor = float64(fac)
